+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Converter for Lithuania (EuroCrops)
 - Converter for Switzerland
+- Converter for Slovenia
 
 ### Fixed